re-organise and clean up modules
