+++ zingolib/src/utils.rs
@@ -1,6 +1,6 @@
 //! General library utilities such as parsing and conversions.
 
-use std::path::Path;
+use std::{path::Path, time::SystemTime};
 
 use tokio::io::AsyncWriteExt as _;
 
@@ -46,6 +46,20 @@ pub(crate) use build_method_push;
 #[cfg(any(test, feature = "test-elevation"))]
 pub(crate) use build_push_list;
 
+/// Writes `bytes` to file at `path`.
+pub(crate) async fn write_to_path(path: &Path, bytes: Vec<u8>) -> std::io::Result<()> {
+    let mut file = tokio::fs::File::create(path).await?;
+    file.write_all(&bytes).await
+}
+
+/// Returns number of seconds since unix epoch.
+pub(crate) fn now() -> u32 {
+    SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .expect("should never fail when comparing with an instant so far in the past")
+        .as_secs() as u32
+}
+
 /// Take a P2PKH taddr and interpret it as a tex addr
 #[cfg(feature = "test-elevation")]
 pub fn interpret_taddr_as_tex_addr(
@@ -57,10 +71,4 @@ pub fn interpret_taddr_as_tex_addr(
         &taddr_bytes,
     )
     .unwrap()
-}
-
-/// Writes `bytes` to file at `path`.
-pub(crate) async fn write_to_path(path: &Path, bytes: Vec<u8>) -> std::io::Result<()> {
-    let mut file = tokio::fs::File::create(path).await?;
-    file.write_all(&bytes).await
 }
